+++ controller_manager/src/ros2_control_node.cpp
@@ -18,6 +18,7 @@
 #include <thread>
 
 #include "controller_manager/controller_manager.hpp"
+#include "rclcpp/executors.hpp"
 #include "realtime_tools/thread_priority.hpp"
 
 using namespace std::chrono_literals;