+++ hardware_interface/include/hardware_interface/actuator_interface.hpp
@@ -228,7 +228,6 @@ public:
    */
   void set_state(const rclcpp_lifecycle::State & new_state) { lifecycle_state_ = new_state; }
 
-protected:
   /// Get the logger of the ActuatorInterface.
   /**
    * \return logger of the ActuatorInterface.
@@ -241,6 +240,13 @@ protected:
    */
   rclcpp::Clock::SharedPtr get_clock() const { return clock_interface_->get_clock(); }
 
+  /// Get the hardware info of the ActuatorInterface.
+  /**
+   * \return hardware info of the ActuatorInterface.
+   */
+  const HardwareInfo & get_hardware_info() const { return info_; }
+
+protected:
   HardwareInfo info_;
   rclcpp_lifecycle::State lifecycle_state_;