+++ hardware_interface/include/hardware_interface/sensor_interface.hpp
@@ -167,7 +167,6 @@ public:
    */
   void set_state(const rclcpp_lifecycle::State & new_state) { lifecycle_state_ = new_state; }
 
-protected:
   /// Get the logger of the SensorInterface.
   /**
    * \return logger of the SensorInterface.
@@ -180,6 +179,13 @@ protected:
    */
   rclcpp::Clock::SharedPtr get_clock() const { return clock_interface_->get_clock(); }
 
+  /// Get the hardware info of the SensorInterface.
+  /**
+   * \return hardware info of the SensorInterface.
+   */
+  const HardwareInfo & get_hardware_info() const { return info_; }
+
+protected:
   HardwareInfo info_;
   rclcpp_lifecycle::State lifecycle_state_;